# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для работы с сервисом расчета генерации электроэнергии ветрогенераторами",
        "title": "Turbines API",
        "contact": {
            "name": "Егор Бляблин"
        },
        "license": {
            "name": "MIT License"
        },
        "version": "0.2.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/api/turbines/": {
            "get": {
                "description": "Возвращает список активных турбин с возможностью фильтрации по названию",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ветрогенераторы"
                ],
                "summary": "Список турбин с фильтрацией",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по названию турбины",
                        "name": "turbineTitle",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список турбин",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ds.Turbine"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Создает новую турбину без изображения",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ветрогенераторы"
                ],
                "summary": "Создание новой турбины",
                "parameters": [
                    {
                        "description": "Данные новой турбины",
                        "name": "turbine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.CreateTurbine"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Созданная турбина",
                        "schema": {
                            "$ref": "#/definitions/ds.Turbine"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/turbines/{turbineId}/": {
            "get": {
                "description": "Возвращает информацию о конкретной турбине по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ветрогенераторы"
                ],
                "summary": "Получение информации о турбине",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID турбины",
                        "name": "turbineId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о турбине",
                        "schema": {
                            "$ref": "#/definitions/ds.Turbine"
                        }
                    },
                    "404": {
                        "description": "Турбина не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Некорректный ID турбины",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Обновляет информацию о турбине по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ветрогенераторы"
                ],
                "summary": "Обновление информации о турбине",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID турбины",
                        "name": "turbineId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "turbine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.UpdateTurbine"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленная турбина",
                        "schema": {
                            "$ref": "#/definitions/ds.Turbine"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Турбина не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Некорректный ID турбины",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Удаляет турбину по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ветрогенераторы"
                ],
                "summary": "Удаление турбины",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID турбины",
                        "name": "turbineId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Турбина успешно удалена"
                    },
                    "404": {
                        "description": "Турбина не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Некорректный ID турбины",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/turbines/{turbineId}/upload-image/": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Добавляет или обновляет изображение для турбины",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ветрогенераторы"
                ],
                "summary": "Добавление изображения к турбине",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID турбины",
                        "name": "turbineId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Файл изображения",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленная турбина",
                        "schema": {
                            "$ref": "#/definitions/ds.Turbine"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Турбина не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Некорректный ID турбины",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ds.CreateTurbine": {
            "type": "object",
            "required": [
                "description",
                "height",
                "is_active",
                "power",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "minLength": 10
                },
                "height": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "power": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                }
            }
        },
...
        "ds.Turbine": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "power": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "ds.UpdateTurbine": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "minLength": 10
                },
                "height": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "power": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "description": "Префикс \"Bearer\" с последующими пробелом и JWT.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}